Confidencialidad en cuadro + Múltiples cambios y correcciones
Antes de extraer las características de video ahora se verifica si realmente el cuadro contiene una cara.

Cambiada la lógica de la extracción de características en video por períodos de tiempo. En caso de tener períodos inválidos por la confidencialidad se aproximan utilizando el anterior y posterior.

Cambios en concatenaArff para unificar correctamente las subpartes al eliminar los silencios en respuestas. Luego de utilizar cada archivo en la concatenación lo borra.

Al obtener las ROI si una zona es inválida porque sus landmarks dan nulos se devuelve una ROI en negro por esa zona.

Reestructuración del método de elimina silencios para simplificarlo y hacerlo más eficiente.

Nuevo archivo donde se tienen todas las funciones codificadas que ya no se utilizan.
